LinkList: refactor to use Stack
