+++ packages/link-list/src/index.tsx
@@ -1,1 +1,3 @@
 export * from './LinkList';
+export { LinkListGroup } from './LinkListGroup';
+export { LinkListItem } from './LinkListItem';
